requirements.txt
flask
flask-cors
tensorflow-cpu==2.20.0
numpy
pandas
scikit-learn
joblib